rename api-server.js to api.js
